Use nan for node 4/5 compatibility
