+++ .travis.yml
@@ -1,9 +1,10 @@
 language: node_js
 node_js:
-    - "0.12"
-    - "0.10"
     - "0.8"
-    - iojs
+    - "0.10"
+    - "0.12"
+    - "4"
+    - "5"
 
 before_install:
