+++ binding.gyp
@@ -1,11 +1,12 @@
 {
-    "targets": [
-	{
-	    "target_name": "syslog",
-	    "sources": [ "syslog.cc" ],
-	    "cflags": [
-		"-fPIC"
-	    ]
-	}
-    ]
+    "targets": [{
+        "target_name": "syslog",
+        "sources": [ "src/node-syslog.cc" ],
+        "cflags": [
+            "-fPIC"
+        ],
+        "include_dirs" : [
+            "<!(node -e \"require('nan')\")"
+        ]
+    }]
 }
