+++ lib/node-syslog.js
@@ -1,4 +1,4 @@
-var SyslogWrapper = require('./build/Release/syslog');
+var SyslogWrapper = require('bindings')('syslog.node');
 
 /*
  * export Syslog as module
