+++ package.json
@@ -1,31 +1,41 @@
 {
-    "name": "node-syslog",
-    "version": "1.2.0",
-    "main": "./node-syslog",
-    "engines": {
-        "node" : ">=0.8.0"
-    },
-    "os": [
-        "!win32"
-    ],
-    "scripts": {
-        "test": "node test.js"
-    },
-    "description": "Node module to support sending messages to syslog daemon",
-    "author": "Nazar Kulyk <nasar.kulyk@googlemail.com>",
-    "homepage": "http://github.com/schamane/node-syslog",
-    "repository" :
+  "name": "node-syslog",
+  "version": "1.2.0",
+  "main": "lib/index",
+  "engines": {
+    "node": ">=0.8.0"
+  },
+  "os": [
+    "!win32"
+  ],
+  "scripts": {
+    "test": "node test"
+  },
+  "description": "Node module to support sending messages to syslog daemon",
+  "author": "Nazar Kulyk <nasar.kulyk@googlemail.com>",
+  "homepage": "http://github.com/schamane/node-syslog",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:schamane/node-syslog.git"
+  },
+  "bugs": {
+    "url": "https://github.com/schamane/node-syslog/issues"
+  },
+  "licenses": [
     {
-        "type" : "git",
-        "url" : "git@github.com:schamane/node-syslog.git"
-    },
-    "bugs": {
-        "url": "https://github.com/schamane/node-syslog/issues"
-    },
-    "licenses": [
-    {
-        "type": "Simplified BSD License",
-        "url": "http://en.wikipedia.org/wiki/BSD_licenses"
-    }],
-    "tags": ["syslog", "system", "log", "logging" ]
+      "type": "Simplified BSD License",
+      "url": "http://en.wikipedia.org/wiki/BSD_licenses"
+    }
+  ],
+  "tags": [
+    "syslog",
+    "system",
+    "log",
+    "logging"
+  ],
+  "dependencies": {
+    "bindings": "^1.2.1",
+    "nan": "^2.1.0"
+  },
+  "gypfile": true
 }
